Feat/um 23 new schema

+++ pages/api/teams/[id]/user.ts
@@ -1,8 +1,10 @@
 import { NextApiRequestQueryBody } from 'interface/api/nextApi';
+import { allowQuery } from 'lib/auth';
+import { UmamiApi } from 'lib/constants';
 import { uuid } from 'lib/crypto';
 import { useAuth } from 'lib/middleware';
 import { NextApiResponse } from 'next';
-import { methodNotAllowed, ok } from 'next-basics';
+import { methodNotAllowed, ok, unauthorized } from 'next-basics';
 import { createTeamUser, deleteTeamUser, getUsersByTeamId } from 'queries';
 
 export interface TeamUserRequestQuery {
@@ -23,12 +25,20 @@ export default async (
   const { id: teamId } = req.query;
 
   if (req.method === 'GET') {
+    if (!(await allowQuery(req, UmamiApi.AuthType.Team))) {
+      return unauthorized(res);
+    }
+
     const user = await getUsersByTeamId({ teamId });
 
     return ok(res, user);
   }
 
   if (req.method === 'POST') {
+    if (!(await allowQuery(req, UmamiApi.AuthType.TeamOwner))) {
+      return unauthorized(res, 'You must be the owner of this team.');
+    }
+
     const { user_id: userId } = req.body;
 
     const updated = await createTeamUser({ id: uuid(), userId, teamId });
@@ -37,6 +47,10 @@ export default async (
   }
 
   if (req.method === 'DELETE') {
+    if (!(await allowQuery(req, UmamiApi.AuthType.TeamOwner))) {
+      return unauthorized(res, 'You must be the owner of this team.');
+    }
+
     const { team_user_id } = req.body;
 
     await deleteTeamUser(team_user_id);
